Final Tic Tac Toe functions
