+++ src/components/TicTacToe/Board.js
@@ -1,27 +1,34 @@
 import React, { PureComponent } from 'react'
 import Square from './Square'
+import { calculateWinner } from '../../containers/Game'
 
 class Board extends PureComponent {
-  constructor(props) {
-    super(props);
+  constructor() {
+    super();
     this.state = {
-      squares: Array(9).fill(0),
+      squares: Array(9).fill(null),
       xIsNext: true,
     };
   }
 
   handleClick(i) {
     const squares = this.state.squares.slice();
-    squares[i] = this.state.xIsNext ? 'X': 'O';
-    this.setState({squares: squares,
-    xIsNext: !this.state.xIsNext,
-  });
-}
+    if (calculateWinner(squares) || squares[i]) {
+      return;
+    }
+    squares[i] = this.state.xIsNext ? 'X' : 'O';
+    this.setState({
+      squares: squares,
+      xIsNext: !this.state.xIsNext,
+    });
+  }
 
   renderSquare(i) {
     return (
-      <Square value={this.state.squares[i]}
-        onClick={() => this.handleClick(i)} />
+      <Square
+        value={this.state.squares[i]}
+        onClick={() => this.handleClick(i)}
+      />
     );
   }
 
@@ -29,17 +36,14 @@ class Board extends PureComponent {
     const winner = calculateWinner(this.state.squares);
     let status;
     if (winner) {
-      status = "Winner: " + winner;
+      status = 'Winner: ' + winner;
     } else {
-      status = "Next player: " + (this.state.xIsNext ? 'X' : 'O');
+      status = 'Next player: ' + (this.state.xIsNext ? 'X' : 'O');
     }
 
-
-
     return (
-      <div className="whole-board">
-        <div className="status"> { status } </div>
-        <div className="grid">
+      <div>
+        <div className="status">{status}</div>
         <div className="board-row">
           {this.renderSquare(0)}
           {this.renderSquare(1)}
@@ -55,30 +59,10 @@ class Board extends PureComponent {
           {this.renderSquare(7)}
           {this.renderSquare(8)}
         </div>
-        </div>
       </div>
     );
   }
 }
 
-function calculateWinner(squares) {
-  const lines = [
-    [0, 1, 2],
-    [3, 4, 5],
-    [6, 7, 8],
-    [0, 3, 6],
-    [1, 4, 7],
-    [2, 5, 8],
-    [0, 4, 8],
-    [2, 4, 6],
-  ];
-  for (let i = 0; i < lines.length; i++) {
-    const [a, b, c] = lines[i];
-    if (squares[a] && squares[a] === squares[b] && squares[a] === squares[c]) {
-      return squares[a];
-    }
-  }
-  return null;
-}
 
   export default Board
