+++ src/containers/Game.js
@@ -7,6 +7,27 @@ import JoinGameDialog from '../components/games/JoinGameDialog'
 import GameTTT from '../components/TicTacToe/GameTTT'
 import './Game.css'
 
+export const calculateWinner = (squares) => {
+  const lines = [
+    [0, 1, 2],
+    [3, 4, 5],
+    [6, 7, 8],
+    [0, 3, 6],
+    [1, 4, 7],
+    [2, 5, 8],
+    [0, 4, 8],
+    [2, 4, 6],
+  ];
+  for (let i = 0; i < lines.length; i++) {
+    const [a, b, c] = lines[i];
+    if (squares[a] && squares[a] === squares[b] && squares[a] === squares[c]) {
+      return squares[a];
+    }
+  }
+  return null;
+}
+
+
 const playerShape = PropTypes.shape({
   userId: PropTypes.string.isRequired,
   name: PropTypes.string
